foreach-ext {foreach}	R Documentation

foreach extension functions

Description

These functions are used to write parallel backends for the foreach package. They should not be used from normal scripts or packages that use the foreach package.

Usage

makeAccum(it)

accumulate(obj, result, tag, ...)

getResult(obj, ...)

getErrorValue(obj, ...)

getErrorIndex(obj, ...)

## S3 method for class 'iforeach'
accumulate(obj, result, tag, ...)

## S3 method for class 'iforeach'
getResult(obj, ...)

## S3 method for class 'iforeach'
getErrorValue(obj, ...)

## S3 method for class 'iforeach'
getErrorIndex(obj, ...)

## S3 method for class 'ixforeach'
accumulate(obj, result, tag, ...)

## S3 method for class 'ixforeach'
getResult(obj, ...)

## S3 method for class 'ixforeach'
getErrorValue(obj, ...)

## S3 method for class 'ixforeach'
getErrorIndex(obj, ...)

## S3 method for class 'ifilteredforeach'
accumulate(obj, result, tag, ...)

## S3 method for class 'ifilteredforeach'
getResult(obj, ...)

## S3 method for class 'ifilteredforeach'
getErrorValue(obj, ...)

## S3 method for class 'ifilteredforeach'
getErrorIndex(obj, ...)

getexports(ex, e, env, good = character(0), bad = character(0))

Arguments

it
foreach iterator.

obj
foreach iterator object.

result
task result to accumulate.

tag
tag of task result to accumulate.

...
unused.

ex
call object to analyze.

e
local environment of the call object.

env
exported environment in which call object will be evaluated.

good
names of symbols that are being exported.

bad
names of symbols that are not being exported.

Note

These functions are likely to change in future versions of the foreach package. When they become more stable, they will be documented.

[Package foreach version 1.5.2 Index]