getDoParWorkers {foreach}	R Documentation

Functions Providing Information on the doPar Backend

Description

The getDoParWorkers function returns the number of execution workers there are in the currently registered doPar backend. It can be useful when determining how to split up the work to be executed in parallel. A 1 is returned by default.

The getDoParRegistered function returns TRUE if a doPar backend has been registered, otherwise FALSE.

The getDoParName function returns the name of the currently registered doPar backend. A NULL is returned if no backend is registered.

The getDoParVersion function returns the version of the currently registered doPar backend. A NULL is returned if no backend is registered.

Usage

getDoParWorkers()

getDoParRegistered()

getDoParName()

getDoParVersion()

Examples

cat(sprintf('%s backend is registered\n',
            if(getDoParRegistered()) 'A' else 'No'))
cat(sprintf('Running with %d worker(s)\n', getDoParWorkers()))
(name <- getDoParName())
(ver <- getDoParVersion())
if (getDoParRegistered())
  cat(sprintf('Currently using %s [%s]\n', name, ver))

[Package foreach version 1.5.2 Index]