getDoSeqWorkers {foreach}	R Documentation

Functions Providing Information on the doSeq Backend

Description

The getDoSeqWorkers function returns the number of execution workers there are in the currently registered doSeq backend. A 1 is returned by default.

The getDoSeqRegistered function returns TRUE if a doSeq backend has been registered, otherwise FALSE.

The getDoSeqName function returns the name of the currently registered doSeq backend. A NULL is returned if no backend is registered.

The getDoSeqVersion function returns the version of the currently registered doSeq backend. A NULL is returned if no backend is registered.

Usage

getDoSeqRegistered()

getDoSeqWorkers()

getDoSeqName()

getDoSeqVersion()

Examples

cat(sprintf('%s backend is registered\n',
            if(getDoSeqRegistered()) 'A' else 'No'))
cat(sprintf('Running with %d worker(s)\n', getDoSeqWorkers()))
(name <- getDoSeqName())
(ver <- getDoSeqVersion())
if (getDoSeqRegistered())
  cat(sprintf('Currently using %s [%s]\n', name, ver))

[Package foreach version 1.5.2 Index]