registerDoSEQ {foreach}	R Documentation

registerDoSEQ

Description

The registerDoSEQ function is used to explicitly register a sequential parallel backend with the foreach package. This will prevent a warning message from being issued if the %dopar% function is called and no parallel backend has been registered.

Usage

registerDoSEQ()

See Also

doParallel::registerDoParallel

Examples

# specify that %dopar% should run sequentially
registerDoSEQ()

[Package foreach version 1.5.2 Index]