setDoPar {foreach}	R Documentation

setDoPar

Description

The setDoPar function is used to register a parallel backend with the foreach package. This isn't normally executed by the user. Instead, packages that provide a parallel backend provide a function named registerDoPar that calls setDoPar using the appropriate arguments.

Usage

setDoPar(fun, data = NULL, info = function(data, item) NULL)

Arguments

fun
A function that implements the functionality of %dopar%.

data
Data to be passed to the registered function.

info
Function that retrieves information about the backend.

See Also

%dopar%

[Package foreach version 1.5.2 Index]