setDoSeq {foreach}	R Documentation

setDoSeq

Description

The setDoSeq function is used to register a sequential backend with the foreach package. This isn't normally executed by the user. Instead, packages that provide a sequential backend provide a function named registerDoSeq that calls setDoSeq using the appropriate arguments.

Usage

setDoSeq(fun, data = NULL, info = function(data, item) NULL)

Arguments

fun
A function that implements the functionality of %dopar%.

data
Data to be passed to the registered function.

info
Function that retrieves information about the backend.

See Also

%dopar%

[Package foreach version 1.5.2 Index]